requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
boto3==1.34.17
botocore==1.34.17
certifi==2023.11.17
charset-normalizer==3.3.2
dataclasses-json==0.6.3
distro==1.9.0
dnspython==2.4.2
exceptiongroup==1.2.0
frozenlist==1.4.1
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.0
langchain-community==0.0.11
langchain-core==0.1.10
langsmith==0.0.80
loguru==0.7.2
marshmallow==3.20.2
momento==1.16.0
momento-wire-types==0.98.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
openai==1.7.1
packaging==23.2
pinecone-client==2.2.4
protobuf==4.25.2
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
slack-bolt==1.18.1
slack_sdk==3.26.2
sniffio==1.3.0
SQLAlchemy==2.0.25
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.0.7
yarl==1.9.4
